Create Keyboard Map

Creates a keyboard map

Usage

create_keyboard_map(scale_table, middle_c_octave = 4)

Arguments

scale_table

an output of create_scale_table

middle_c_octave

octave number for middle C. There are varying conventions for the octave number for middle C. The default for this function is 4, but other software can use 3 or even some other number

Value

the keyboard map. This is a data.table with eight columns:

note_number: the MIDI note number from .NN_MIN through .NN_MAX

key_name: note name of the key in 12EDO. This is the key you'd normally press to play this note number

key_octave: the octave the key normally plays

note_name: the note name from the scale table

ratio_frac: ratio for the note as a vulgar fraction

degree: the scale degree of the note

octave: the octave number of the note

freq: the frequency in Hz

cents: cents above default MIDI note .NN_MIN, which has frequency .FREQ_MIN.

ref_keyname: some synthesizers, including the Korg Minilogue XD, let you retune a key as an offset in cents from another reference key. This column is the name of that reference key.

ref_octave: the octave number of the reference key

ref_offset: the offset in cents from the reference key

Details

The function is currently hard-coded to compute the map so that middle C with frequency .FREQ_MIDDLE_Cis mapped to MIDI note number .NN_MIDDLE_C and scale degree 0. With the current constants this is the same as it is on 12EDO with A440 on note 69. This note is 6000 cents above MIDI note number 0.

Examples

if (FALSE) {

hexany_harmonics <- c(1, 3, 5, 7)
hexany_choose = 2
hexany <- create_scale_table(hexany_harmonics, hexany_choose)
print(hexany_keyboard_map <- create_keyboard_map(hexany))
dekany_harmonics <- c(1, 7, 9, 11, 13)
dekany_choose <- 2
dekany <- create_scale_table(dekany_harmonics, dekany_choose)
print(dekany_keyboard_map <- create_keyboard_map(dekany))
print(
  vanilla_keyboard_map <- create_keyboard_map(create_12edo_scale_table())
)
print(
  eikosany_keyboard_map_c3 <-
    create_keyboard_map(create_scale_table(), middle_c_octave = 3)
)
}