Create Keyboard Map

Creates a keyboard map

Usage

keyboard_map(scale_table, middle_c_octave = 4)

Arguments

scale_table

a scale table from ps_scale_table, cps_scale_table, or et_scale_table

middle_c_octave

octave number for middle C. There are varying conventions for the octave number for middle C. The default for this function is 4, but other software can use 3 or even some other number

Value

the keyboard map. This is a data.table with eight columns:

note_number: the MIDI note number from .NN_MIN through .NN_MAX

key_name: note name of the key in 12EDO. This is the key you'd normally press to play this note number

key_octave: the octave the key normally plays

note_name: the note name from the scale table

ratio_frac: ratio for the note as a vulgar fraction

degree: the scale degree of the note

octave: the octave number of the note

freq: the frequency in Hz

cents: cents above default MIDI note .NN_MIN, which has frequency .FREQ_MIN.

ref_keyname: some synthesizers, including the Korg Minilogue XD, let you retune a key as an offset in cents from another reference key. This column is the name of that reference key.

ref_octave: the octave number of the reference key

ref_offset: the offset in cents from the reference key

Details

The function is currently hard-coded to compute the map so that middle C with frequency .FREQ_MIDDLE_Cis mapped to MIDI note number .NN_MIDDLE_C and scale degree 0. With the current constants this is the same as it is on 12EDO with A440 on note 69. This note is 6000 cents above MIDI note number 0.

Normally you would only use this to remap a keyboard to a scale with more than 12 notes per octave. For scales with 12 or fewer notes to the octave, it's easier to remap all octaves using the offsets provided in the scale table.

Examples


eikosany <- cps_scale_table()
print(eikosany_keyboard_map <- keyboard_map(eikosany))
#>      note_number key_name key_octave note_name ratio_frac degree octave
#>   1:           0        C         -1     1x3x5          1      0      1
#>   2:           1       C#         -1    5x9x11      33/32      1      1
#>   3:           2        D         -1     1x7x9      21/20      2      1
#>   4:           3       D#         -1    1x3x11      11/10      3      1
#>   5:           4        E         -1     3x5x9        9/8      4      1
#>  ---                                                                   
#> 124:         123       D#          9    1x3x11      11/10      3      7
#> 125:         124        E          9     3x5x9        9/8      4      7
#> 126:         125        F          9     1x5x7        7/6      5      7
#> 127:         126       F#          9    3x9x11      99/80      6      7
#> 128:         127        G          9    1x7x11      77/60      7      7
#>            freq     cents ref_keyname ref_octave ref_offsets
#>   1:   32.70320  2400.000           C          2           0
#>   2:   33.72517  2453.273          C#          2         -47
#>   3:   34.33836  2484.467          C#          2         -16
#>   4:   35.97352  2565.004           D          2         -35
#>   5:   36.79110  2603.910           D          2           4
#>  ---                                                        
#> 124: 2302.30497  9765.004           D          8         -35
#> 125: 2354.63009  9803.910           D          8           4
#> 126: 2441.83861  9866.871          D#          8         -33
#> 127: 2590.09310  9968.914           E          8         -31
#> 128: 2686.02247 10031.875           E          8          32

# 12-EDO for sanity check
print(vanilla_keyboard_map <- keyboard_map(et_scale_table()))
#>      note_number key_name key_octave note_name ratio_frac degree octave
#>   1:           0        C         -1         C          1      0     -1
#>   2:           1       C#         -1     C#|Db  1461/1379      1     -1
#>   3:           2        D         -1         D  1714/1527      2     -1
#>   4:           3       D#         -1     D#|Eb  1785/1501      3     -1
#>   5:           4        E         -1         E    635/504      4     -1
#>  ---                                                                   
#> 124:         123       D#          9     D#|Eb  1785/1501      3      9
#> 125:         124        E          9         E    635/504      4      9
#> 126:         125        F          9         F  3249/2434      5      9
#> 127:         126       F#          9     F#|Gb   1393/985      6      9
#> 128:         127        G          9         G  2213/1477      7      9
#>              freq cents ref_keyname ref_octave ref_offsets
#>   1:     8.175799     0           C          0           0
#>   2:     8.661957   100          C#          0           0
#>   3:     9.177024   200           D          0           0
#>   4:     9.722718   300          D#          0           0
#>   5:    10.300861   400           E          0           0
#>  ---                                                      
#> 124:  9956.063479 12300          D#         10           0
#> 125: 10548.081821 12400           E         10           0
#> 126: 11175.303406 12500           F         10           0
#> 127: 11839.821527 12600          F#         10           0
#> 128: 12543.853951 12700           G         10           0

# check middle C setting
print(
  eikosany_keyboard_map_c3 <-
    keyboard_map(cps_scale_table(), middle_c_octave = 3))
#>      note_number key_name key_octave note_name ratio_frac degree octave
#>   1:           0        C         -2     1x3x5          1      0      0
#>   2:           1       C#         -2    5x9x11      33/32      1      0
#>   3:           2        D         -2     1x7x9      21/20      2      0
#>   4:           3       D#         -2    1x3x11      11/10      3      0
#>   5:           4        E         -2     3x5x9        9/8      4      0
#>  ---                                                                   
#> 124:         123       D#          8    1x3x11      11/10      3      6
#> 125:         124        E          8     3x5x9        9/8      4      6
#> 126:         125        F          8     1x5x7        7/6      5      6
#> 127:         126       F#          8    3x9x11      99/80      6      6
#> 128:         127        G          8    1x7x11      77/60      7      6
#>            freq     cents ref_keyname ref_octave ref_offsets
#>   1:   32.70320  2400.000           C          2           0
#>   2:   33.72517  2453.273          C#          2         -47
#>   3:   34.33836  2484.467          C#          2         -16
#>   4:   35.97352  2565.004           D          2         -35
#>   5:   36.79110  2603.910           D          2           4
#>  ---                                                        
#> 124: 2302.30497  9765.004           D          8         -35
#> 125: 2354.63009  9803.910           D          8           4
#> 126: 2441.83861  9866.871          D#          8         -33
#> 127: 2590.09310  9968.914           E          8         -31
#> 128: 2686.02247 10031.875           E          8          32