Create Product Set Scale Table

Creates a scale table from a product set definition

Usage

ps_scale_table(
  ps_def = c("1x3x5", "5x9x11", "1x7x9", "1x3x11", "3x5x9", "1x5x7", "3x9x11", "1x7x11",
    "5x7x9", "3x5x11", "1x3x7", "7x9x11", "1x5x9", "3x7x9", "5x7x11", "1x9x11", "3x5x7",
    "1x3x9", "1x5x11", "3x7x11")
)

Arguments

ps_def

the product set scale definition. This is a character vector of products. Each product is a set of any number of integers separated by a lower-case "x". For example, the ps_def of the 1-3-5-7 Hexany is

c("1x3", "1x5", "1x7", "3x5", "3x7", "5x7")

The default is the ps_def for the 1-3-5-7-9-11 Eikosany.

Value

a data.table with six columns:

note_name: the given product set definition, re-ordered by the degrees of the resulting scale (character)

ratio: the ratio that defines the note, as a number between 1 and 2

ratio_frac: the ratio as a vulgar fraction (character)

ratio_cents: the ratio in cents (hundredths of a semitone)

interval_cents: interval between this note and the previous note

degree: scale degree from zero to (number of notes) - 1

See also

Examples


# the default yields the 1-3-5-7-9-11 Eikosany
print(eikosany <- ps_scale_table())
#>     note_name    ratio ratio_frac ratio_cents interval_cents degree
#>  1:     1x3x5 1.000000          1     0.00000             NA      0
#>  2:    5x9x11 1.031250      33/32    53.27294       53.27294      1
#>  3:     1x7x9 1.050000      21/20    84.46719       31.19425      2
#>  4:    1x3x11 1.100000      11/10   165.00423       80.53704      3
#>  5:     3x5x9 1.125000        9/8   203.91000       38.90577      4
#>  6:     1x5x7 1.166667        7/6   266.87091       62.96090      5
#>  7:    3x9x11 1.237500      99/80   368.91423      102.04332      6
#>  8:    1x7x11 1.283333      77/60   431.87513       62.96090      7
#>  9:     5x7x9 1.312500      21/16   470.78091       38.90577      8
#> 10:    3x5x11 1.375000       11/8   551.31794       80.53704      9
#> 11:     1x3x7 1.400000        7/5   582.51219       31.19425     10
#> 12:    7x9x11 1.443750    231/160   635.78514       53.27294     11
#> 13:     1x5x9 1.500000        3/2   701.95500       66.16987     12
#> 14:     3x7x9 1.575000      63/40   786.42219       84.46719     13
#> 15:    5x7x11 1.604167      77/48   818.18885       31.76665     14
#> 16:    1x9x11 1.650000      33/20   866.95923       48.77038     15
#> 17:     3x5x7 1.750000        7/4   968.82591      101.86668     16
#> 18:     1x3x9 1.800000        9/5  1017.59629       48.77038     17
#> 19:    1x5x11 1.833333       11/6  1049.36294       31.76665     18
#> 20:    3x7x11 1.925000      77/40  1133.83013       84.46719     19
#> 21:    1x3x5' 2.000000          2  1200.00000       66.16987     20
#>     note_name    ratio ratio_frac ratio_cents interval_cents degree

# Kraig Grady's Eikosany as two complementary extended Dekanies
# See _Microtonality and the Tuning Systems of Erv Wilson_, pages 127 - 131
# for the process used to create these scales
print(grady_a <- ps_scale_table(c(
  "1x3x11",
  "1x9",
  "3x9x11",
  "1x7x11",
  "1x3x7",
  "7x9x11",
  "3x7x9",
  "1x9x11",
  "1x3x9",
  "1x7",
  "3x7x11",
  "1x7x9"
)))
#>     note_name    ratio ratio_frac ratio_cents interval_cents degree
#>  1:       1x7 1.000000          1     0.00000             NA      0
#>  2:    3x7x11 1.031250      33/32    53.27294       53.27294      1
#>  3:     1x7x9 1.125000        9/8   203.91000      150.63706      2
#>  4:    1x3x11 1.178571      33/28   284.44704       80.53704      3
#>  5:       1x9 1.285714        9/7   435.08410      150.63706      4
#>  6:    3x9x11 1.325893    297/224   488.35704       53.27294      5
#>  7:    1x7x11 1.375000       11/8   551.31794       62.96090      6
#>  8:     1x3x7 1.500000        3/2   701.95500      150.63706      7
#>  9:    7x9x11 1.546875      99/64   755.22794       53.27294      8
#> 10:     3x7x9 1.687500      27/16   905.86500      150.63706      9
#> 11:    1x9x11 1.767857      99/56   986.40204       80.53704     10
#> 12:     1x3x9 1.928571      27/14  1137.03910      150.63706     11
#> 13:      1x7' 2.000000          2  1200.00000       62.96090     12
print(grady_a_offsets <- offset_matrix(grady_a))
#>           C   C#    D   D#    E    F   F#    G   G#    A    A#     B
#> 1x7       0 -100 -200 -300 -400 -500 -600 -700 -800 -900 -1000 -1100
#> 3x7x11   53  -47 -147 -247 -347 -447 -547 -647 -747 -847  -947 -1047
#> 1x7x9   204  104    4  -96 -196 -296 -396 -496 -596 -696  -796  -896
#> 1x3x11  284  184   84  -16 -116 -216 -316 -416 -516 -616  -716  -816
#> 1x9     435  335  235  135   35  -65 -165 -265 -365 -465  -565  -665
#> 3x9x11  488  388  288  188   88  -12 -112 -212 -312 -412  -512  -612
#> 1x7x11  551  451  351  251  151   51  -49 -149 -249 -349  -449  -549
#> 1x3x7   702  602  502  402  302  202  102    2  -98 -198  -298  -398
#> 7x9x11  755  655  555  455  355  255  155   55  -45 -145  -245  -345
#> 3x7x9   906  806  706  606  506  406  306  206  106    6   -94  -194
#> 1x9x11  986  886  786  686  586  486  386  286  186   86   -14  -114
#> 1x3x9  1137 1037  937  837  737  637  537  437  337  237   137    37
print(grady_b <- ps_scale_table(c(
  "3x5x11",
  "1x5x9",
  "3x5x9x11",
  "5x7x11",
  "3x5x7",
  "1x5x11",
  "1x3x5",
  "5x9x11",
  "3x5x9",
  "1x5x7",
  "3x5x7x11",
  "5x7x9"
)))
#>     note_name    ratio ratio_frac ratio_cents interval_cents degree
#>  1:     1x3x5 1.000000          1     0.00000             NA      0
#>  2:    5x9x11 1.031250      33/32    53.27294       53.27294      1
#>  3:     3x5x9 1.125000        9/8   203.91000      150.63706      2
#>  4:     1x5x7 1.166667        7/6   266.87091       62.96090      3
#>  5:  3x5x7x11 1.203125      77/64   320.14385       53.27294      4
#>  6:     5x7x9 1.312500      21/16   470.78091      150.63706      5
#>  7:    3x5x11 1.375000       11/8   551.31794       80.53704      6
#>  8:     1x5x9 1.500000        3/2   701.95500      150.63706      7
#>  9:  3x5x9x11 1.546875      99/64   755.22794       53.27294      8
#> 10:    5x7x11 1.604167      77/48   818.18885       62.96090      9
#> 11:     3x5x7 1.750000        7/4   968.82591      150.63706     10
#> 12:    1x5x11 1.833333       11/6  1049.36294       80.53704     11
#> 13:    1x3x5' 2.000000          2  1200.00000      150.63706     12
print(grady_b_offsets <- offset_matrix(grady_b))
#>             C   C#    D   D#    E    F   F#    G   G#    A    A#     B
#> 1x3x5       0 -100 -200 -300 -400 -500 -600 -700 -800 -900 -1000 -1100
#> 5x9x11     53  -47 -147 -247 -347 -447 -547 -647 -747 -847  -947 -1047
#> 3x5x9     204  104    4  -96 -196 -296 -396 -496 -596 -696  -796  -896
#> 1x5x7     267  167   67  -33 -133 -233 -333 -433 -533 -633  -733  -833
#> 3x5x7x11  320  220  120   20  -80 -180 -280 -380 -480 -580  -680  -780
#> 5x7x9     471  371  271  171   71  -29 -129 -229 -329 -429  -529  -629
#> 3x5x11    551  451  351  251  151   51  -49 -149 -249 -349  -449  -549
#> 1x5x9     702  602  502  402  302  202  102    2  -98 -198  -298  -398
#> 3x5x9x11  755  655  555  455  355  255  155   55  -45 -145  -245  -345
#> 5x7x11    818  718  618  518  418  318  218  118   18  -82  -182  -282
#> 3x5x7     969  869  769  669  569  469  369  269  169   69   -31  -131
#> 1x5x11   1049  949  849  749  649  549  449  349  249  149    49   -51